first cut at MacOS install

+++ bootstrap.sh
@@ -3,12 +3,23 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
-if [ "`hostnamectl | grep Debian`" != "" ]; then
-echo ${bold}
-echo ============================
-echo upgrade debian to buster
-echo ============================
-echo ${normal}
+if [[ "$OSTYPE" =~ darwin* ]]; then
+  echo ${bold}
+  echo ============================
+  echo prepare MacOS
+  echo ============================
+  echo ${normal}
+  sudo xcode-select --install
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  brew doctor
+  xargs -a <(awk '! /^ *(#|$)/' "brewrequirements.txt") -r -- brew install
+  xargs -a <(awk '! /^ *(#|$)/' "caskrequirements.txt") -r -- brew cask install
+elif [ "`hostnamectl | grep Debian`" != "" ]; then
+  echo ${bold}
+  echo ============================
+  echo upgrade debian to buster
+  echo ============================
+  echo ${normal}
   sudo sed -i s/stretch/buster/g /etc/apt/sources.list
   sudo sed -i s/stretch/buster/g /etc/apt/sources.list.d/cros.list
   sudo sed -i s/stretch/buster/g /etc/apt/sources.list.d/cros-gpu.list
@@ -26,20 +37,21 @@ elif [ -v SOMMELIER_VERSION ]; then
 fi
 
 # Install base apt packages
-echo ${bold}
-echo ============================
-echo installing base apt packages
-echo ============================
-echo ${normal}
-xargs -a <(awk '! /^ *(#|$)/' "aptrequirements.txt") -r -- sudo apt -y install
+if ! [[ "$OSTYPE" =~ darwin* ]]; then
+  echo ${bold}
+  echo ============================
+  echo installing base apt packages
+  echo ============================
+  echo ${normal}
+  xargs -a <(awk '! /^ *(#|$)/' "aptrequirements.txt") -r -- sudo apt -y install
 
-
-echo ${bold}
-echo =====================================
-echo ensure nvim is our default vim editor
-echo =====================================
-echo ${normal}
-sudo update-alternatives --set vim /usr/bin/nvim
+  echo ${bold}
+  echo =====================================
+  echo ensure nvim is our default vim editor
+  echo =====================================
+  echo ${normal}
+  sudo update-alternatives --set vim /usr/bin/nvim
+fi
 
 # Install pyenv.  zshrc already setup to run it.
 echo ${bold}
@@ -72,6 +84,16 @@ echo =================================
 echo installing latest python for user
 echo =================================
 echo ${normal}
+
+if ! [[ "$OSTYPE" =~ darwin* ]]; then
+  export LDFLAGS="${LDFLAGS} -L/usr/local/opt/zlib/lib"
+  export CPPFLAGS="${CPPFLAGS} -I/usrlocal/opt/zlib/include"
+  export PKG_CONFIG_PATH="${PKG_CONFIG_PATH} /usr/local/opt/zlib/lib/pkgconfig"
+
+  export LDFLAGS="${LDFLAGS} -L/usr/local/opt/sqllite/lib"
+  export CPPFLAGS="${CPPFLAGS} -I/usrlocal/opt/sqllite/include"
+  export PKG_CONFIG_PATH="${PKG_CONFIG_PATH} /usr/local/opt/sqllite/lib/pkgconfig"
+fi
 pyenv latest install 2 -s
 pyenv latest install 3 -s
 pyenv latest global
@@ -84,7 +106,10 @@ echo installing base python3 packages
 echo ================================
 echo ${normal}
 python3 -m pip install --upgrade pip
-python3 -m pip install --upgrade -r requirements.txt
+
+if ! [[ "$OSTYPE" =~ darwin* ]]; then
+  python3 -m pip install --upgrade -r requirements.txt
+fi
 pyenv activate neovim3
 python3 -m pip install --upgrade pip
 python3 -m pip install --upgrade -r neovim-requirements.txt
@@ -148,46 +173,51 @@ else
   echo nerd-fonts already exists
 fi
 
-echo ${bold}
-echo =======================================================
-echo installing latest hyper terminal from https://hyper.is/
-echo =======================================================
-echo ${normal}
-PKG_OK=$(dpkg-query -W --showformat='${Status}\n' hyper | grep "install ok installed" )
-if [ "" == "$PKG_OK" ]; then
-  wget https://releases.hyper.is/download/deb -O hyper.deb
-  sudo apt -y install ./hyper.deb
-  rm hyper.deb
-else
-  echo ${bold}already installed${normal}
+
+if ! [[ "$OSTYPE" =~ darwin* ]]; then
+  echo ${bold}
+  echo =======================================================
+  echo installing latest hyper terminal from https://hyper.is/
+  echo =======================================================
+  echo ${normal}
+
+  PKG_OK=$(dpkg-query -W --showformat='${Status}\n' hyper | grep "install ok installed" )
+  if [ "" == "$PKG_OK" ]; then
+    wget https://releases.hyper.is/download/deb -O hyper.deb
+    sudo apt -y install ./hyper.deb
+    rm hyper.deb
+  else
+    echo ${bold}already installed${normal}
+  fi
+  echo ${bold}
+  echo ================================================
+  echo installing gitrocket into hyper terminal plugins
+  echo ================================================
+  echo ${normal}
+  if [ ! -d ~/.config/hyper/.hyper_plugins/node_modules/gitrocket ]; then
+    mkdir -p ~/.config/hyper/.hyper_plugins/node_modules
+    npm install --prefix ~/.config/hyper/.hyper_plugins gitrocket
+  fi
 fi
 
-echo ${bold}
-echo ================================================
-echo installing gitrocket into hyper terminal plugins
-echo ================================================
-echo ${normal}
-if [ ! -d ~/.config/hyper/.hyper_plugins/node_modules/gitrocket ]; then
-  mkdir -p ~/.config/hyper/.hyper_plugins/node_modules
-  npm install --prefix ~/.config/hyper/.hyper_plugins gitrocket
-fi
-
-echo ${bold}
-echo =================
-echo installing vscode
-echo =================
-echo ${normal}
-if [ ! -f /etc/apt/trusted.gpg.d/microsoft.gpg ]; then
-  curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
-  sudo install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
-fi
-if [ ! -f /etc/apt/sources.list.d/vscode.list  ]; then
-  sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
-  sudo apt update
-fi
-PKG_OK=$(dpkg-query -W --showformat='${Status}\n' code | grep "install ok installed" )
-if [ "" == "$PKG_OK" ]; then
-  sudo apt -y install code
+if ! [[ "$OSTYPE" =~ darwin* ]]; then
+  echo ${bold}
+  echo =================
+  echo installing vscode
+  echo =================
+  echo ${normal}
+  if [ ! -f /etc/apt/trusted.gpg.d/microsoft.gpg ]; then
+    curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
+    sudo install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
+  fi
+  if [ ! -f /etc/apt/sources.list.d/vscode.list  ]; then
+    sudo sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
+    sudo apt update
+  fi
+  PKG_OK=$(dpkg-query -W --showformat='${Status}\n' code | grep "install ok installed" )
+  if [ "" == "$PKG_OK" ]; then
+    sudo apt -y install code
+  fi
 fi
 
 echo ${bold}
